requirements.txt
Django>=5.1.3
knock-subdomains>=7.0.2